apps/api/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
fastapi[all]
firebase-admin
python-dotenv
numpy
torch
torchvision
opencv-python

